# Use `dotnet run --no-build`