@t0yv0 2022-09-08
On a project with `pulumi new aws-csharp`.

Before:

```
Benchmark 1: pulumi preview
  Time (mean ± σ):      7.715 s ±  0.327 s    [User: 4.959 s, System: 1.328 s]
  Range (min … max):    7.035 s …  8.109 s    10 runs
 ```

After:

```
Benchmark 1: pulumi preview
  Time (mean ± σ):      6.625 s ±  0.300 s    [User: 3.826 s, System: 1.158 s]
  Range (min … max):    6.278 s …  7.280 s    10 runs
```

About:

```
running 'dotnet build -nologo .'
  Determining projects to restore...

  All projects are up-to-date for restore.

  how-does-dotnet-do-it-2 -> /Users/anton/tmp/how-does-dotnet-do-it-2/bin/Debug/net6.0/how-does-dotnet-do-it-2.dll



Build succeeded.
    0 Warning(s)
    0 Error(s)

Time Elapsed 00:00:00.87

'dotnet build -nologo .' completed successfully
CLI          
Version      3.39.2
Go Version   go1.19
Go Compiler  gc

Plugins
NAME    VERSION
aws     5.13.0
dotnet  unknown

Host     
OS       darwin
Version  12.5.1
Arch     x86_64

This project is written in dotnet: executable='/Users/anton/.nix-profile/bin/dotnet' version='6.0.300'

Current Stack: dev

Found no resources associated with dev

Found no pending operations associated with dev

Backend        
Name           pulumi.com
URL            https://app.pulumi.com/t0yv0
User           t0yv0
Organizations  t0yv0, pulumi

Dependencies:
NAME        VERSION
Pulumi      3.39.3
Pulumi.Aws  5.13.0

Pulumi locates its logs in /var/folders/mb/s1jx27m10t3082fqclpgytt80000gn/T/ by default
```