@t0yv0 2022-09-08
.NET folks any concerns with this? Is this perhaps a new flag in .NET 6 (I could not find which version it was introduced in), should we hedge?