@Zaid-Ajaj 2022-09-08
I don't know about --no-build but AFAIK if you just `dotnet run` without flags _and_ without `dotnet build` first, it will check whether anything has changes and rebuild if needed but I need to double check this myself.