@t0yv0 2022-09-08
Yeah it does - but it takes a while to do that, while we almost certainly run `dotnet build` right before it every single time; hence the change saves 1s.