@justinvp 2022-09-08
> .NET folks any concerns with this?

~1s savings sounds great. I can't think of any downsides.

> Is this perhaps a new flag in .NET 6 (I could not find which version it was introduced in)

I just confirmed the `--no-build` flag is available in .NET Core 3.1. And from looking at the source code, it's been available for a long time: https://github.com/dotnet/sdk/blame/43c13f2f5330149581dbf3ad44c22fdbd3db6ddf/src/dotnet/commands/dotnet-run/RunCommandParser.cs#L36